Add reusable console log library and update apps
