+++ apps/app1/app-data/app1.json
@@ -139,9 +139,15 @@
       "width": 200,
       "height": 150
     },
+    "console-log": {
+      "title": "Console Logs",
+      "content": "<div class=\"console-log-window\" style=\"width:100%;height:100%;background:#000;color:#0f0;font-family:monospace;overflow:auto;font-size:16px;padding:4px;\"></div>",
+      "width": 400,
+      "height": 250
+    },
     "html-studio": {
       "title": "HTML Studio",
-      "content": "<div id=\"htmlstudio-container\" style=\"width:100%;height:100%;overflow:hidden;\"></div>",
+      "content": "<iframe src=\"html-studio.html\" style=\"width:100%;height:100%;border:none;\"></iframe>",
       "width": 600,
       "height": 450
     },
