+++ apps/app1/app-js/cloud-storage.js
@@ -1,0 +1,1 @@
+WinAPI.createWindow('cloud-storage');
